Moves BookListing to components
